Refactor webhooks and return status from send notification
