--- src/server/controllers/webhook-events.ts
@@ -1,228 +1,0 @@
-import gql from 'graphql-tag';
-
-import {StatusToPRQuery} from '../../types/gql-types';
-import {github} from '../../utils/github';
-import {sendNotification} from '../controllers/notifications';
-import {pullRequestsModel} from '../models/pullRequestsModel';
-import {userModel} from '../models/userModel';
-
-type User = {
-  login: string;
-};
-
-type ReviewHook = {
-  state: string, user: User;
-};
-
-type PullRequest = {
-  title: string; user: User; html_url: string;
-};
-
-type RepositoryHook = {
-  name: string;
-};
-
-type Reviewer = {
-  login: string;
-};
-
-type PullRequestHook = {
-  action: string; pull_request: PullRequest; repository: RepositoryHook;
-  requested_reviewer: Reviewer;
-};
-
-type PullRequestReviewHook = {
-  action: string; review: ReviewHook; pull_request: PullRequest;
-  repository: RepositoryHook;
-};
-
-type StatusHook = {
-  sha: string; name: string; state: 'error' | 'failure' | 'pending' | 'success';
-  description: string;
-  repository: RepositoryHook;
-  commit: {author: User;}
-};
-
-// Triggered when the status of a Git commit changes.
-export async function handleStatus(hookData: StatusHook): Promise<boolean> {
-  if (hookData.state !== 'error' && hookData.state !== 'failure') {
-    return false;
-  }
-
-  // There was an error that should be reported to the PR Owner
-  const author = hookData.commit.author;
-  const loginDetails = await userModel.getLoginDetails(author.login);
-  if (!loginDetails) {
-    // Commit author isn't logged in so not token to figure out
-    // appropriate PR's affect by this commit status
-    return false;
-  }
-
-  const repo = hookData.repository;
-  const statusPR = await github().query<StatusToPRQuery>({
-    query: statusToPR,
-    variables: {
-      query: `type:pr repo:${hookData.name} ${hookData.sha}`,
-    },
-    fetchPolicy: 'network-only',
-    // We use the commit author's token for this request
-    context: {token: loginDetails.githubToken}
-  });
-
-  if (!statusPR.data.pullRequests || !statusPR.data.pullRequests.nodes) {
-    return false;
-  }
-
-  let sentNotifications = false;
-  for (const prData of statusPR.data.pullRequests.nodes) {
-    if (!prData || prData.__typename !== 'PullRequest') {
-      continue;
-    }
-
-    // We'll want to message the author
-    if (!prData.author) {
-      continue;
-    }
-
-    // If we can't confirm that the PR commit === the status commit - return;
-    if (!prData.commits.nodes || prData.commits.nodes.length === 0) {
-      continue;
-    }
-
-    // Ensure the commit exists
-    const commitNode = prData.commits.nodes[0];
-    if (!commitNode) {
-      continue;
-    }
-
-
-    const commit = commitNode.commit;
-    if (commit.oid === hookData.sha) {
-      const commitDetails =
-          await pullRequestsModel.getCommitDetails(prData.id, commit.oid);
-
-      if (!commitDetails || commitDetails.status !== hookData.state) {
-        sentNotifications = true;
-
-        await pullRequestsModel.setCommitStatus(
-            prData.id, commit.oid, hookData.state);
-
-        await sendNotification(prData.author.login, {
-          title: hookData.description,
-          body: `[${repo.name}] ${prData.title}`,
-          requireInteraction: false,
-          icon: '/images/notification-images/icon-192x192.png',
-          data: {
-            url: prData.url,
-          }
-        });
-      }
-    }
-  }
-
-  return sentNotifications;
-}
-
-// Triggered when a pull request is assigned, unassigned, labeled, unlabeled,
-// opened, edited, closed, reopened, or synchronized
-export async function handlePullRequest(hookBody: PullRequestHook):
-    Promise<boolean> {
-  if (hookBody.action !== 'review_requested') {
-    return false;
-  }
-
-  const pullRequest = hookBody.pull_request;
-  const repo = hookBody.repository;
-  const user = pullRequest.user;
-  const reviewer = hookBody.requested_reviewer;
-
-  // The reviewer's dash will have a new entry under incoming
-  await userModel.markUserForUpdate(reviewer.login);
-  // The PR owner's dash should update with an incoming entry
-  await userModel.markUserForUpdate(user.login);
-
-  const notification = {
-    title: `${user.login} requested a review`,
-    body: `[${repo.name}] ${pullRequest.title}`,
-    requireInteraction: true,
-    icon: '/images/notification-images/icon-192x192.png',
-    data: {
-      url: pullRequest.html_url,
-    }
-  };
-
-  await sendNotification(reviewer.login, notification);
-
-  return true;
-}
-
-export async function handlePullRequestReview(hookData: PullRequestReviewHook):
-    Promise<boolean> {
-  if (hookData.action !== 'submitted') {
-    return false;
-  }
-
-  const review = hookData.review;
-  const repo = hookData.repository;
-  const pullReq = hookData.pull_request;
-
-  let notificationTitle = null;
-
-  // The author of the review should change to no longer having any action.
-  await userModel.markUserForUpdate(review.user.login);
-  // The PR owner should update with latest info.
-  await userModel.markUserForUpdate(pullReq.user.login);
-
-  if (review.state === 'approved') {
-    notificationTitle = `${review.user.login} approved your PR`;
-  } else if (review.state === 'changes_requested') {
-    notificationTitle = `${review.user.login} requested changes`;
-  } else if (review.state === 'commented') {
-    // Check if the PR author is the commenter
-    if (review.user.login !== pullReq.user.login) {
-      notificationTitle = `${review.user.login} commented on your PR`;
-    }
-  }
-
-  if (notificationTitle) {
-    await sendNotification(pullReq.user.login, {
-      title: notificationTitle,
-      body: `[${repo.name}] ${pullReq.title}`,
-      requireInteraction: true,
-      icon: '/images/notification-images/icon-192x192.png',
-      data: {
-        url: pullReq.html_url,
-      },
-    });
-  }
-
-  return true;
-}
-
-const statusToPR = gql`
-query StatusToPR($query: String!) {
-  pullRequests: search (
-    first: 10,
-    type: ISSUE,
-    query: $query
-  ) {
-    nodes {
-      ... on PullRequest {
-        id,
-        title,
-        url,
-        author {
-          login
-        },
-        commits(last: 1) {
-          nodes {
-            commit {
-              oid,
-            }
-          }
-        }
-      }
-    }
-  }
-}
-`;
